// desktopApp/src-ui/src/api/types.ts
export interface ApiResult<T = unknown> {
  code: number
  message: string
  data: T
}